requirements.txt
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
coverage==5.4
Django==3.1.4
djangorestframework==3.12.2
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
pytest-django==4.1.0
pytz==2020.4
redis==3.5.3
requests==2.25.1
sqlparse==0.4.1
tld==0.12.5
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
zipp==3.4.0
